backend/requirements.txt
fastapi==0.115.6
httpx==0.27.2
python-dotenv==1.0.1
uvicorn[standard]==0.34.0
tokenc==0.1.0